EmotionFAD_Clean/requirements.txt
# Core dependencies
Flask==3.0.2
Flask-SocketIO==5.3.5
Werkzeug==3.0.3
Jinja2==3.1.3
itsdangerous==2.2.0
click==8.1.7

# Web server
gunicorn==21.2.0
eventlet==0.33.3

# Computer Vision
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
fer==0.1.0

# WebSockets
python-engineio==4.8.2
python-socketio==5.10.0

# Environment
python-dotenv==1.0.1

# System dependencies (included for completeness)
setuptools==69.0.3
wheel==0.42.0

# Ensure these are explicitly included
gevent==23.7.0
gevent-websocket==0.10.1
